atualizacao da home com ultimos saldos e contas

+++ src/components/ContasBox.js
@@ -6,33 +6,39 @@ import {View,
 import Svg, {Polygon, Rect} from 'react-native-svg'
 import SvgPath from './SvgPath'
 
+const editar = (edit, editDell) => {
+    if(editDell == true){
+        return <View style={style.editDellBox}>
+                    <TouchableOpacity onPress={edit}>
+                        <Svg width="40" height="40" viewBox="0 0 540.329 540.329">
+                            <Polygon points={`0.002,540.329 58.797,532.66 7.664,481.528`}/>
+                            <Polygon points={`16.685,412.341 10.657,458.56 81.765,529.668 127.983,523.64 442.637,208.992 331.338,97.688`}/>
+                            <SvgPath type="edit"/>                        
+                        </Svg>
+                    </TouchableOpacity>
+                    <TouchableOpacity>
+                        <Svg width="40" height="40" viewBox="0 0 512 512">
+                            <Polygon points={`353.574,176.526 313.496,175.056 304.807,412.34 344.885,413.804`}/>
+                            <Rect x={`235.948`} y={`175.791`} width={"40.104"} height={"237.28"}/>
+                            <Polygon points={`207.186,412.334 198.497,175.049 158.419,176.52 167.109,413.804`}/>
+                            <SvgPath type="delete"/>
+                        </Svg>
+                    </TouchableOpacity>
+                </View>
+    }
+}
+
 const ContasBox = (props) => {
-    const {edit, show} = props;
+    const {edit, show, editDell} = props;
     return (
         <View style={style.contasBox} >
             <TouchableOpacity style={style.containContasBox} onPress={show}>
                 <View>
                     <Text style={style.texto}>Nome: Pagar aluguel</Text>
-                    <Text style={style.texto}>Valor: 400,00 R$</Text>
+                    <Text style={style.texto}>Valor: R$ 400,00</Text>
                 </View>
             </TouchableOpacity>
-            <View style={style.editDellBox}>
-                <TouchableOpacity onPress={edit}>
-                    <Svg width="40" height="40" viewBox="0 0 540.329 540.329">
-                        <Polygon points={`0.002,540.329 58.797,532.66 7.664,481.528`}/>
-                        <Polygon points={`16.685,412.341 10.657,458.56 81.765,529.668 127.983,523.64 442.637,208.992 331.338,97.688`}/>
-                        <SvgPath type="edit"/>                        
-                    </Svg>
-                </TouchableOpacity>
-                <TouchableOpacity>
-                    <Svg width="40" height="40" viewBox="0 0 512 512">
-                        <Polygon points={`353.574,176.526 313.496,175.056 304.807,412.34 344.885,413.804`}/>
-                        <Rect x={`235.948`} y={`175.791`} width={"40.104"} height={"237.28"}/>
-                        <Polygon points={`207.186,412.334 198.497,175.049 158.419,176.52 167.109,413.804`}/>
-                        <SvgPath type="delete"/>
-                    </Svg>
-                </TouchableOpacity>
-            </View>
+            {editar(edit, editDell)}
         </View>
     )
 }
